requirements.txt
# RadioTrack - Maine Department of Corrections Radio Inventory Management System
# Copyright (c) 2025 Arthur Belanger (github.com/coding-with-arty)
# All rights reserved.

# Requirements for Production Deployment
# Updated: January 2025

# Core Web Framework
streamlit==1.40.2

# Data Processing & Analysis
pandas==2.2.0
plotly==5.18.0

# Image Processing
Pillow==10.1.0

# PDF Generation
reportlab==4.0.7

# Excel File Processing
openpyxl==3.1.2

# Environment Management
python-dotenv==1.0.0

# Security & Authentication
bcrypt==4.1.1
pyotp==2.9.0
cryptography==41.0.7

# System Monitoring & Scheduling
schedule==1.2.0
psutil==5.9.6

# HTTP Requests (for future API integrations)
requests==2.31.0

# Application Directory Management
appdirs==1.4.4

# Optional: Email notifications (uncomment if needed)
# smtplib is built-in, but for enhanced email features:
# secure-smtplib==0.1.1

# Development dependencies (consider moving to dev-requirements.txt for production)
# pytest==7.4.0
# black==23.7.0
# flake8==6.0.0